(NOOP) Rename submodule using new class name.
